Move name/rfn check to googlefonts profile
Also add rationale and make it a **FAIL** as it is a strong requirement for Google Fonts that families do not use a "Reserved Font Name"
(issue #2779)
